Cleaned up my workspace and began trying PGN data. I decided to postpone working on documentation to try my hand at implementing PGN data into my code, as well as trying Chess960, but Chess960 would have needed me to greatly refactor my code for some edge cases, and documenting this entire project would be a mission in itself. I'm writing this commit message before I go to bed and hopefully I'll open a new branch with development on Chess960, as well as working in a separate repository on the documentation I can then migrate over to this repository.
Thanks to everyone who helped me. Really couldn't be here without everyone who encouraged me along the way.

Also, I heard from Matt, the creator of Heimdall, that a simple negamax algorithm with alpha-beta pruning to filter out unreasonable moves, and a transposition table to store recently searched positions, would already put an engine into about the 2000 elo level. But of course, that's for another day. I'll probably do Chess960 as an additional challenge before I start making an engine.

But then again, who knows?
