+++ Perft/Perft.cs
@@ -1,8 +1,29 @@
 using System.Diagnostics;
 using Chess.MoveGen;
+using Newtonsoft.Json;
+using System.Text;
 
 namespace Chess.Perft
 {
+    public struct GameData
+    {
+        public string FEN { get; set; }
+        public Dictionary<int, ulong> Depths { get; set; }
+
+        public override string ToString() => $"<GameData FEN = {FEN}, depths = {string.Join(", ", Depths)}>";
+    }
+
+    public struct UnitTestResult
+    {
+        public string FEN { get; set; }
+        public int Depth { get; set; }
+        public ulong ExpectedNodes { get; set; }
+        public ulong? ReturnedNodes { get; set; }
+        public bool PassedTest { get { return ExpectedNodes == ReturnedNodes; } }
+        public double? TimeTaken { get; set; }
+        public string? ErrorMessage { get; set; }
+    }
+
     public static class Perft
     {
         public static ulong BasePerftTest(Board board, int depth, bool bulk = true)
@@ -126,6 +147,86 @@ namespace Chess.Perft
 
                 Console.WriteLine($"Depth: {depth}  |  Nodes: {count}  |  Time taken: {sw.ElapsedMilliseconds}ms");
             }
+        }
+
+        static void RunTests(int maxDepth, int[]? blockPositions = null, int startAt = 0)
+        {
+            blockPositions ??= [];
+
+            Console.WriteLine("Loading tests...");
+
+            FileStream fs = new("perft.json", FileMode.Open, FileAccess.Read);
+            byte[] buffer = new byte[fs.Length];
+            fs.Read(buffer);
+
+            string content = Encoding.UTF8.GetString(buffer);
+
+            GameData[]? json = JsonConvert.DeserializeObject<GameData[]>(content) ?? throw new Exception("JSON was null.");
+
+            Console.WriteLine("Finished loading tests.");
+
+            int correct = 0;
+            int incorrect = 0;
+            int errors = 0;
+
+            List<UnitTestResult> unitTestResults = [];
+
+            // --------------------------------------------------
+
+            for (int i = startAt; i < json.Length; i++)
+            {
+                if (blockPositions.Contains(i + 1)) continue;
+                
+                GameData entry = json[i];
+
+                Board board = new(entry.FEN);
+
+                Stopwatch sw = new();
+                
+                for (int depth = 1; depth < maxDepth + 1; depth++) // search from depth 1 to 4 (efficiency)
+                {
+                    Console.Clear();
+                    Console.WriteLine($"Position {i + 1}:\n----------{new string('-', $"{i + 1}".Length)}\nFEN: {entry.FEN}\nBoard:\n{board}\n\n");
+
+                    sw.Reset();
+                    ulong expected = entry.Depths[depth - 1];
+
+                    sw.Start();
+                    ulong returned = Perft.BasePerftTest(board, depth);
+                    sw.Stop();
+
+                    if (expected != returned)
+                    {
+                        throw new Exception($"incorrect amount returned - expected {expected} and got back {returned}. [Failed depth {depth}]");
+                    }
+
+                    double time = (double)sw.ElapsedMilliseconds / 1000;
+
+                    unitTestResults.Add(
+                        new UnitTestResult()
+                        {
+                            FEN = entry.FEN,
+                            Depth = depth,
+                            ExpectedNodes = expected,
+                            ReturnedNodes = returned,
+                            TimeTaken = time,
+                            ErrorMessage = null
+                        }
+                    );
+
+                    if (returned == expected) correct++; else incorrect++;
+                }
+            }
+
+            Console.WriteLine($"Finished tests!\n\nCorrect: {correct}\nIncorrect: {incorrect}\nErrors: {errors}\n");
+
+            Console.WriteLine("Uploading to JSON file...");
+
+            string JSONresults = JsonConvert.SerializeObject(unitTestResults, Formatting.Indented);
+
+            File.WriteAllText("results.json", JSONresults);
+
+            Console.WriteLine("Finished writing to JSON file!");
         }
     }
 }
